moved profile pic into component
